Move `urlpath` from `manio.py` to `k8s.py`.
